+++ tests/test_infer/test_ops/triton/test_context_attn_unpad.py
@@ -2,7 +2,7 @@ import pytest
 import torch
 from packaging import version
 
-from colossalai.inference.flash_decoding_utils import get_alibi_slopes
+from colossalai.inference.modeling.models.nopadding_baichuan import get_alibi_slopes
 from colossalai.kernel.triton import context_attention_unpadded
 from colossalai.utils import get_current_device
 from tests.test_infer.test_ops.triton.kernel_utils import generate_caches_and_block_tables_v2, torch_attn_ref
